prop-settings: find home for useSettings
